tennis_lessen_scraper/requirements.txt
python-dotenv>=1.0.0
playwright>=1.40.0
openpyxl>=3.1.0
cryptography>=42.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
